performance(inspector) Repeated Number Input Keypresses - Changed a test from a browser test to a jest test.
