+++ editor/src/uuiui/inputs/number-input.spec.browser2.tsx
@@ -79,89 +79,4 @@ describe('NumberInput', () => {
     )
     checkPlaceholder(result, 'Mixed')
   })
-  it('triggers multiple transient updates and a single non-transient one when holding a key down and eventually releasing it', async () => {
-    const storeHookForTest = getStoreHook()
-    let value = cssNumber(100)
-    let transientUpdates: Array<UnknownOrEmptyInput<CSSNumber>> = []
-
-    function updateValue(number: UnknownOrEmptyInput<CSSNumber>): void {
-      if (isUnknownInputValue(number) || isEmptyInputValue(number)) {
-        throw new Error(`Unexpected value: ${number}`)
-      } else {
-        value = number
-      }
-      reRender()
-    }
-
-    function onTransientSubmitValue(number: UnknownOrEmptyInput<CSSNumber>): void {
-      transientUpdates.push(number)
-      updateValue(number)
-    }
-    let nonTransientUpdates: Array<UnknownOrEmptyInput<CSSNumber>> = []
-    function onSubmitValue(number: UnknownOrEmptyInput<CSSNumber>): void {
-      nonTransientUpdates.push(number)
-      updateValue(number)
-    }
-
-    function reRender() {
-      return render(
-        <TestInspectorContextProvider
-          selectedViews={storeHookForTest.getState().editor.selectedViews}
-          editorStoreData={storeHookForTest}
-        >
-          <NumberInput
-            testId='keydowntest'
-            controlStatus='simple'
-            value={value}
-            numberType={'AnyValid'}
-            defaultUnitToHide={null}
-            // eslint-disable-next-line react/jsx-no-bind
-            onTransientSubmitValue={onTransientSubmitValue}
-            // eslint-disable-next-line react/jsx-no-bind
-            onSubmitValue={onSubmitValue}
-          />
-        </TestInspectorContextProvider>,
-      )
-    }
-
-    const result = reRender()
-    const inputElement = result.getByTestId('keydowntest')
-    for (let count = 0; count < 10; count++) {
-      act(() => {
-        fireEvent(
-          inputElement,
-          new KeyboardEvent('keydown', {
-            bubbles: true,
-            cancelable: true,
-            key: 'ArrowUp',
-            keyCode: keycode('ArrowUp'),
-          }),
-        )
-        fireEvent(
-          inputElement,
-          new KeyboardEvent('keyup', {
-            bubbles: true,
-            cancelable: true,
-            key: 'ArrowUp',
-            keyCode: keycode('ArrowUp'),
-          }),
-        )
-      })
-    }
-    // Need this for the delay that is in the `NumberInput`.
-    await wait(700)
-    expect(transientUpdates).toEqual([
-      cssNumber(101),
-      cssNumber(102),
-      cssNumber(103),
-      cssNumber(104),
-      cssNumber(105),
-      cssNumber(106),
-      cssNumber(107),
-      cssNumber(108),
-      cssNumber(109),
-      cssNumber(110),
-    ])
-    expect(nonTransientUpdates).toEqual([cssNumber(110)])
-  })
 })
